breakup (and remove) firebaseService
